// src/enums.ts
export enum DisciplineType {
    RUNNING = "RUNNING",
    JUMPING = "JUMPING",
    THROWING = "THROWING",
    COMBINED_EVENTS = "COMBINED_EVENTS"
}

export enum AgeGroup {
    CHILD = "CHILD",
    YOUTH = "YOUTH",
    JUNIOR = "JUNIOR",
    ADULT = "ADULT",
    SENIOR = "SENIOR",
    NOT_ELIGIBLE = "NOT_ELIGIBLE"
}

export enum Gender {
    MALE = "MALE",
    FEMALE= "FEMALE",
    OTHER = "OTHER"
}

export enum ResultType {
    POINTS = "POINTS",
    TIME = "TIME",
    DISTANCE = "DISTANCE"
}